pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ImageVisionAI"
version = "0.0.1"
description = "Real-Time Image Classification Web Application with FastAPI and Gradio"
authors = ["SkxPhan <110777139+SkxPhan@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
gunicorn = "^22.0.0"
SQLAlchemy = "^2.0.31"
sqlalchemy-utils = "^0.41.2"
pydantic = "^2.8.2"
python-jose = "^3.3.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-cov = "^5.0.0"
tox = "^4.16.0"
flake8 = "^7.1.0"
mypy = "^1.11.0"
black = "^24.4.2"
isort = "^5.13.2"
bandit = "^1.7.9"
flake8-pyproject = "^1.2.3"
pre-commit = "^3.8.0"

[tool.pytest.ini_options]
addopts = "--cov=src"
testpaths = ["tests"]
markers = [
    "unit: mark a test as a unit test",
    "integration: mark a test as an integration test",
    "api: mark a test as an api test"
]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.flake8]
max-line-length = 80

[tool.black]
line-length = 80
